Extract Kenbokukai scraper module
