+++ tests/test_kenbokukai_scraper.py
@@ -1,9 +1,9 @@
 import unittest
 
-from scrape_kendo_schedule import (
-    dedupe_events,
+from kendo_keiko.scrapers.kenbokukai import (
     parse_kenbokukai_events_from_text,
 )
+from scrape_kendo_schedule import dedupe_events
 
 
 class KenbokukaiParserTests(unittest.TestCase):
